refactor: use wrapper for instance storage to make it easier to replace localstorage with other storage
