+++ packages/renderer-worker/src/parts/SaveState/SaveState.js
@@ -1,31 +1,14 @@
-import * as LocalStorage from '../LocalStorage/LocalStorage.js'
+import * as InstanceStorage from '../InstanceStorage/InstanceStorage.js'
 import * as Preferences from '../Preferences/Preferences.js'
 import * as RendererProcess from '../RendererProcess/RendererProcess.js'
+import * as SerializeViewlet from '../SerializeViewlet/SerializeViewlet.js'
 import * as SessionStorage from '../SessionStorage/SessionStorage.js'
 import * as ViewletStates from '../ViewletStates/ViewletStates.js'
 import * as Workspace from '../Workspace/Workspace.js'
 
-const serializeInstance = (instance) => {
-  if (instance && instance.factory && instance.factory.saveState) {
-    return instance.factory.saveState(instance.state)
-  }
-  return undefined
-}
-
-const serializeInstances = (instances) => {
-  const serialized = Object.create(null)
-  for (const value of Object.values(instances)) {
-    const serializedInstance = serializeInstance(value)
-    if (serializedInstance) {
-      serialized[value.moduleId] = serializedInstance
-    }
-  }
-  return serialized
-}
-
 const getStateToSave = () => {
   const instances = ViewletStates.getAllInstances()
-  const savedInstances = serializeInstances(instances)
+  const savedInstances = SerializeViewlet.serializeInstances(instances)
   return {
     instances: savedInstances,
     mainEditors: [],
@@ -40,8 +23,8 @@ const getStateToSave = () => {
 
 export const saveViewletState = async (id) => {
   const instance = ViewletStates.getInstance(id)
-  const savedState = serializeInstance(instance)
-  await LocalStorage.setJson(id, savedState)
+  const savedState = SerializeViewlet.serializeInstance(instance)
+  await InstanceStorage.setJson(id, savedState)
   if (instance.factory.saveChildState) {
     const childIds = instance.factory.saveChildState(instance.state)
     await Promise.all(childIds.map(saveViewletState))
@@ -55,8 +38,8 @@ export const handleVisibilityChange = async (visibilityState) => {
   if (visibilityState === 'hidden') {
     const stateToSave = getStateToSave()
     await Promise.all([
-      LocalStorage.setJson('workspace', stateToSave.workspace),
-      LocalStorage.setJsonObjects(stateToSave.instances),
+      InstanceStorage.setJson('workspace', stateToSave.workspace),
+      InstanceStorage.setJsonObjects(stateToSave.instances),
       SessionStorage.setJson('workspace', stateToSave.workspace),
     ])
     // console.log('[renderer worker] state was saved')
@@ -72,13 +55,6 @@ export const hydrate = async () => {
   await RendererProcess.invoke('Window.onVisibilityChange')
 }
 
-export const getSavedState = () => {
-  if (Workspace.isTest()) {
-    return undefined
-  }
-  return LocalStorage.getJson('stateToSave')
-}
-
 export const getSavedViewletState = (viewletId) => {
-  return LocalStorage.getJson(viewletId)
+  return InstanceStorage.getJson(viewletId)
 }
